Add basic test setup and a test using the Pythia converter
The test setup is designed with the other converters already in mind,
such that adding tests for those should be as easy as defining inputs.

The comparison tests look at the GenParticle, Muon, Photon, Electron,
EFlowTrack, EFlowPhoton, EFlowNeutralHadron and Jet collections of
delphes and checks that we have the same contents in the edm4hep output.
For each delphes candidate and edm4hep candidate only the kinematics are
compared at the moment to check whether the two are the same. For all
ReconstructedParticles in the "global" collection we also check whether
they have the same relations to the MCParticles as the corresponding
delphes particle.